reorganised files removed DebugTool class
