+++ code/libs/CoSpaceFunctions.cpp
