+++ code/libs/CoSpaceFunctions.hpp
