+++ code/libs/ColorRecognition.cpp
